requirements.txt
langchain
langchain-community
langchain-openai
langgraph
chromadb
python-dotenv
tiktoken
openai